refactor: fixed optional type and following dependency
