+++ src/ai/backend/client/cli/admin/scaling_group.py
@@ -2,14 +2,13 @@ import sys
 
 import click
 
-from ai.backend.cli.types import ExitCode
+from ai.backend.cli.params import BoolExprType, JSONParamType, OptionalType
+from ai.backend.cli.types import ExitCode, Undefined, undefined
 from ai.backend.client.func.scaling_group import _default_detail_fields, _default_list_fields
 from ai.backend.client.output.fields import scaling_group_fields
 from ai.backend.client.session import Session
 
-from ...types import Undefined, undefined
 from ..extensions import pass_ctx_obj
-from ..params import BoolExprType, JSONParamType, OptionalType
 from ..types import CLIContext
 from . import admin
