+++ src/ai/backend/client/func/keypair_resource_policy.py
@@ -1,9 +1,10 @@
 from typing import Any, Iterable, Sequence
 
+from ...cli.types import Undefined, undefined
 from ..output.fields import keypair_resource_policy_fields
 from ..output.types import FieldSpec
 from ..session import api_session
-from ..types import Undefined, set_if_set, undefined
+from ..types import set_if_set
 from .base import BaseFunction, api_function, resolve_fields
 
 __all__ = ("KeypairResourcePolicy",)
