+++ src/ai/backend/client/types.py
@@ -1,42 +1,8 @@
 from __future__ import annotations
 
-import enum
 from typing import Any, Callable, MutableMapping
 
-__all__ = (
-    "Sentinel",
-    "Undefined",
-    "sentinel",
-    "undefined",
-)
-
-
-class Sentinel(enum.Enum):
-    """
-    A special type to represent a special value to indicate closing/shutdown of queues.
-    """
-
-    TOKEN = 0
-
-    def __bool__(self) -> bool:
-        # It should be evaluated as False when used as a boolean expr.
-        return False
-
-
-class Undefined(enum.Enum):
-    """
-    A special type to represent an undefined value.
-    """
-
-    TOKEN = 0
-
-    def __bool__(self) -> bool:
-        # It should be evaluated as False when used as a boolean expr.
-        return False
-
-
-sentinel = Sentinel.TOKEN
-undefined = Undefined.TOKEN
+from ai.backend.cli.types import Undefined, undefined
 
 
 def set_if_set(
